backend/atom/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.1
attrs==23.1.0
bcrypt==4.0.1
boto3==1.26.72
botocore==1.29.72
CacheControl==0.12.11
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==38.0.4
et-xmlfile==1.1.0
firebase-admin==6.1.0
flasgger==0.9.5
Flask==2.1.1
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==4.11.1
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.19.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
influxdb==5.3.1
influxdb-client==1.30.0
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.0.1
mistune==2.0.5
msgpack==1.0.5
mysqlclient==2.0.3
ncclient==0.6.10
netaddr==0.8.0
netmiko==3.4.0
ntc-templates==3.3.0
numpy==1.24.3
openpyxl==3.0.7
pandas==1.5.3
paramiko==3.1.0
ping3==4.0.4
plogger==1.0.7
ply==3.11
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.7.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pyspnego==0.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-nmap==0.7.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywinos==1.2.7
pywinrm==0.4.3
PyYAML==6.0
requests==2.26.0
requests-credssp==2.0.0
requests-ntlm==1.2.0
rsa==4.9
Rx==3.2.0
s3transfer==0.6.1
scp==0.14.5
six==1.16.0
SQLAlchemy==1.3.23
tenacity==8.2.2
textfsm==1.1.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
waitress==2.0.0
Werkzeug==2.0.0
xmltodict==0.12.0
zipp==3.15.0